Fixed a few warnings

+++ src/detail/sorting.cpp
@@ -244,7 +244,7 @@ void here_counting_sort
 )
 noexcept
 {
-	const std::size_t size = std::distance(begin,end);
+	const std::size_t size = static_cast<std::size_t>(std::distance(begin,end));
 	if (incr) {
 		lal::detail::sorting::counting_sort<T,lal::detail::sorting::non_decreasing_t>
 		(begin, end, n, size, key);
@@ -344,7 +344,7 @@ err_type exe_rand_sorting(const std::string& option, std::ifstream& fin) noexcep
 		fin >> R >> s >> n;
 		auto this_sort = [&](Ui_it begin, Ui_it end) -> void {
 			lal::detail::sorting::bit_sort<Ui>
-			(begin, end, std::distance(begin, end));
+			(begin, end, static_cast<std::size_t>(std::distance(begin, end)));
 		};
 		for (Ui k = 0; k < R; ++k) {
 			const err_type e = check_sorting("bit", s, n, this_sort, true);
@@ -362,7 +362,7 @@ err_type exe_rand_sorting(const std::string& option, std::ifstream& fin) noexcep
 		std::vector<char> seen(n, 0);
 		auto bsm = [&](Ui_it begin, Ui_it end) -> void {
 			lal::detail::sorting::bit_sort_mem<Ui>
-			(begin, end, std::distance(begin, end), &seen[0]);
+			(begin, end, static_cast<std::size_t>(std::distance(begin, end)), &seen[0]);
 		};
 		// execute test
 		for (Ui k = 0; k < R; ++k) {
